feat: added MVC

+++ public/debug/debugger_database.php
@@ -9,7 +9,10 @@
     <h1>Database Debugger</h1>
 
     <?php 
-    require_once "../connect.php";
+    if ($_SERVER['REMOTE_ADDR'] !== '127.0.0.1') {
+        die("Access denied");
+    }
+    require_once "../../config/connect.php";
 
     if (isset($_POST['ResetDB'])) {
         $conn->query("DROP DATABASE IF EXISTS HMS");
